+++ index.html
@@ -7,10 +7,7 @@
   </head>
 
   <body>
-    <div class="loading">
-      <p>gimme a sec</p>
-    </div>
-    <div id="test"></div>
+    <div id='root'></div>
   </body>
-  <script src="bundle.js"></script>
+  <script src="bundle-v1.0.js"></script>
 </html>
