+++ js/index.js
@@ -1,0 +1,7 @@
+import ReactDOM from 'react-dom';
+import App from './App'
+
+ReactDOM.render(
+  App(),
+  document.getElementById("root")
+);
