+++ webpack.config.js
@@ -1,8 +1,8 @@
 module.exports = {
-  entry: "./js/main.js",
+  entry: "./js/index.js",
   output: {
     path: __dirname,
-    filename: "bundle.js"
+    filename: "bundle-v1.0.js",
   },
   module: {
     rules: [
@@ -11,9 +11,12 @@ module.exports = {
         exclude: /(node_modules)/,
         loader: 'babel-loader',
         query: {
-          presets: ['@babel/preset-env']
+          presets: ['@babel/preset-env', '@babel/preset-react']
         }
       }
     ]
-  }
+  },
+  devServer: {
+    historyApiFallback: true,
+  },
 };
